requirements.txt
numpy
torch
azureml-core
matplotlib